+++ CHANGELOG.md
@@ -46,7 +46,9 @@
 * Compatible to GNU grep 3.8
 * Don't use external pwd command anymore
 * Doesn't hang anymore when there's no local resolver
-* Added --mtls feature to support client authentication 
+* Added --mtls feature to support client authentication
+* Added list of signature_algorithms used for client authentication
+* Added support for signature_algorithms_cert extension
 
 
 ### Features implemented / improvements in 3.0